Modify hierachy
