+++ va_rust_utils/Cargo.toml
@@ -15,6 +15,10 @@ regex = "1.11.1"
 
 [profile.release]
 opt-level = 3
+lto = true
+target-cpu = "native"
 
 [profile.dev]
 opt-level = 3
+lto = true
+target-cpu = "native"
